Fix require statements
