+++ lib/geordi/commands/setup_firefox_for_selenium.rb
@@ -1,6 +1,6 @@
 desc 'setup_firefox_for_selenium', '[sic]'
 def setup_firefox_for_selenium(version)
-  require File.expand_path('../../firefox_for_selenium', __FILE__)
+  require 'geordi/firefox_for_selenium'
 
   Geordi::FirefoxForSelenium.install(version)
 end
