+++ lib/geordi/commands/test.rb
@@ -1,9 +1,9 @@
 desc 'test', 'Run all employed tests'
 def test
-  invoke :with_rake
-  invoke :unit
-  invoke :rspec
-  invoke :cucumber
+  invoke 'with_rake'
+  invoke 'test_unit'
+  invoke 'rspec'
+  invoke 'cucumber'
 
   success 'Successfully ran tests.'
 end
