+++ lib/geordi/commands/version.rb
@@ -1,0 +1,5 @@
+desc 'version', 'Print the current version of geordi'
+def version
+  require 'geordi/version'
+  puts 'Geordi ' + Geordi::VERSION
+end
