+++ lib/geordi/commands/vnc_show.rb
@@ -1,6 +1,6 @@
 desc 'vnc_show', 'Show the hidden VNC window'
 def vnc_show
-  require File.expand_path('../../cucumber', __FILE__)
+  require 'geordi/cucumber'
 
   Geordi::Cucumber.new.launch_vnc_viewer
 end
