+++ lib/geordi/cucumber.rb
@@ -1,5 +1,5 @@
 require "rubygems"
-require File.join(File.dirname(__FILE__), 'firefox_for_selenium')
+require 'geordi/firefox_for_selenium'
 require 'tempfile'
 
 module Geordi
@@ -203,7 +203,7 @@ module Geordi
         VNC_ENV_VARIABLES.each do |variable|
           ENV["OUTER_#{variable}"] = ENV[variable] if ENV[variable]
         end
-        ENV["BROWSER"] = ENV["LAUNCHY_BROWSER"] = File.expand_path(File.join(File.dirname(__FILE__), '../../bin/launchy_browser'))
+        ENV["BROWSER"] = ENV["LAUNCHY_BROWSER"] = File.expand_path('../../../bin/launchy_browser', __FILE__)) # FIXME
         ENV["DISPLAY"] = VNC_DISPLAY
 
         note 'Selenium is running in a VNC window. Use cuc-show to view it.'
